refactor: input view 로직은 view 디렉토리에서 관리
