+++ src/web/view/PurchaseForm.js
