+++ src/web/view/WinningNumberInputs.js
