+++ src/web/winningNumbers/WinningNumbers.js
@@ -1,10 +1,10 @@
 import './winningNumbers.css';
 import { createElement } from '../utils/dom';
 import WinningNumberHeaders from './WinningNumberHeaders';
-import WinningNumberInputs from './WinningNumberInputs';
+import WinningNumberInputs from '../view/WinningNumberInputs';
 
 export default function WinningNumbers(playLotto) {
-  const winningNumberContainer = document.createElement('div', { class: 'winning-number-container' });
+  const winningNumberContainer = createElement('div', { class: 'winning-number-container' });
 
   WinningNumberHeaders(winningNumberContainer);
   const { winningNumbersArray, bonusNumber } = WinningNumberInputs(winningNumberContainer);
